move anything related to analytical PASTIS to separate folder
