+++ pastis/analytical_pastis/aperture_definition.py
